Splited HomePage code into some smaller reusable components
