+++ snortly/src/pages/PostPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function PostPage() {
+    return (
+        <div>PostPage</div>
+    )
+}
+
+export default PostPage
